+++ project/game_components/text/text.h
@@ -8,7 +8,7 @@
 
 /* Predef */
 
-struct Sprite create_text(const char *const text, const SDL_Color inner_color, const SDL_Color outer_color, const struct Vec2 screen_pos, const unsigned int size, const unsigned int border_thickness, int *const exit_code);
+struct Sprite create_text(const char *const text, const SDL_Color inner_color, const SDL_Color outer_color, const struct Vec2 screen_pos, const unsigned short size, const unsigned short border_thickness, int *const exit_code);
 
 
 #endif /// TEXT_H
